\name{get.DerivedFrom}
\alias{get.DerivedFrom}
\alias{get.DerivedFrom-method}
\alias{get.DerivedFrom,ANY-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{get.DerivedFrom}
\description{Return the derived from string from the object (gives the "parent mir"")}
\usage{
get.DerivedFrom(theObject)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{theObject}{object returned from one of the getMirs_Metatdata functions}
}
\details{
Return the derived from string from the object (gives the "parent mir"")
}
\value{
Return the derived from string from the object (gives the "parent mir"")
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{#See examples in tests directory in the install.}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
