\name{get.GeneId}
\alias{get.GeneId}
\alias{get.GeneId-method}
\alias{get.GeneId,ANY-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{get.GeneId}
\description{return the id (numeric digits) of a gene as a string}
\usage{
get.GeneId(theObject)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{theObject}{object returned from one of the getMetadata_Gene functions}
}
\details{return the id (numeric digits) of a gene as a string}
\value{return the id (numeric digits) of a gene as a string}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{#See examples in tests directory in the install.}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
