\name{get.GeneStructure}
\alias{get.GeneStructure}
\alias{get.GeneStructure-method}
\alias{get.GeneStructure,ANY-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{get.GeneStructure}
\description{
returns information about the gene structure at location probed
}
\usage{
get.GeneStructure(theObject)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{theObject}{object returned from one of the getMetadata_Probe functions}
}
\details{
returns information about the gene structure at location probed - a vector of strings, with the names() being the genes to which the structure applies
}
\value{
returns information about the gene structure at location probed - a vector of strings, with the names() being the genes to which the structure applies
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{#See examples in tests directory in the install.}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
