\name{get.LocationEnd}
\alias{get.LocationEnd}
\alias{get.LocationEnd-method}
\alias{get.LocationEnd,ANY-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{get.LocationEnd}
\description{Get the genomic end location as a numeric}
\usage{
get.LocationEnd(theObject)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{theObject}{object returned from one of the getMetadata_Gene functions}
}
\details{Get the genomic end location as a numeric}
\value{Get the genomic end location as a numeric}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{#See examples in tests directory in the install.}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
