\name{get.MirType}
\alias{get.MirType}
\alias{get.MirType-method}
\alias{get.MirType,ANY-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{get.MirType}
\description{
Return the mir type string from the object
}
\usage{
get.MirType(theObject)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{theObject}{object returned from one of the getMirs_Metatdata functions}
}
\details{
Return the mir type string from the object
}
\value{
Return the mir type string from the object
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{#See examples in tests directory in the install.}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
