\name{get.Name}
\alias{get.Name}
\alias{get.Name-method}
\alias{get.Name,ANY-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{get.Name}
\description{return the name of a probe}
\usage{
get.Name(theObject)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{theObject}{object returned from one of the getMetadata_Probe functions}
}
\details{
return the name of a probe as a string
}
\value{
return the name of a probe as a string
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{#See examples in tests directory in the install.}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
