\name{get.SampleAnnotations}
\alias{get.SampleAnnotations}
\alias{get.SampleAnnotations-method}
\alias{get.SampleAnnotations,ANY-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{get.SampleAnnotations}
\description{
returns a data.frame with four columns named: sample, patient, diseaseCode, diseaseString, typeCode, and typeString
}
\usage{
get.SampleAnnotations(theObject)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{theObject}{An object returned from one of the getDataObject_ functions.}
}
\details{
returns a data.frame with four columns named: sample, patient, diseaseCode, diseaseString, typeCode, and typeString
}
\value{
returns a data.frame with four columns named: sample, patient, diseaseCode, diseaseString, typeCode, and typeString
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{#See examples in tests directory in the install.}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
