\name{get.VersionIndex}
\alias{get.VersionIndex}
\alias{get.VersionIndex-method}
\alias{get.VersionIndex,ANY-method}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{get.VersionIndex}
\description{Get the version index as a string. This takes the form of XofY (1of2, 2of2, etc.) indicating multiple entries in the source data for this gene.}
\usage{
get.VersionIndex(theObject)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{theObject}{object returned from one of the getMetadata_Gene functions}
}
\details{Get the version index as a string. This takes the form of XofY (1of2, 2of2, etc.) indicating multiple entries in the source data for this gene.}
\value{Get the version index as a string. This takes the form of XofY (1of2, 2of2, etc.) indicating multiple entries in the source data for this gene.}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{#See examples in tests directory in the install.}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
